Renamed cuda to cupy in kernels datastructure
