+++ src/qttools/kernels/datastructure/cupy/_cupy_rawkernel.py
@@ -4,7 +4,7 @@ import cupy as cp
 import numpy as np
 
 from qttools import NDArray
-from qttools.kernels.datastructure.cuda import THREADS_PER_BLOCK
+from qttools.kernels.datastructure.cupy import THREADS_PER_BLOCK
 
 _reduction = cp.RawKernel(
     r"""
